django-project/requirements/base.txt
# ─── Core ────────────────────────────────────────────────
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
drf-yasg==1.21.8
django-redis==5.4.0

# ─── Celery & Queue ──────────────────────────────────────
celery==5.4.0
redis==5.0.8
kombu==5.4.2

# ─── Database ────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Storage & Media ────────────────────────────────────
Pillow==10.4.0

# ─── Config ──────────────────────────────────────────────
python-decouple==3.8

# ─── Phone Validation ────────────────────────────────────
django-phonenumber-field==8.0.0
phonenumbers==8.13.40

# ─── Monitoring ──────────────────────────────────────────
sentry-sdk==2.8.0

# ─── Server ──────────────────────────────────────────────
gunicorn==22.0.0
gevent==24.2.1

# ─── SMS (Optional) ─────────────────────────────────────
# kavenegar==1.1.3
